Move storybook to root
