+++ .storybook/preview.css.ts
@@ -1,5 +1,5 @@
 import { globalStyle } from '@vanilla-extract/css';
-import { vars } from '../src/styles/theme.css';
+import { vars } from '../packages/components/src/styles/theme.css';
 
 globalStyle('.dark .docs-story', {
   background: vars.color.background
