+++ package.json
@@ -11,6 +11,7 @@
     "ci": "npm install && npm run generate && npm run package && npm run type && (npm run lint:ci || true) && npm run test:ci",
     "generate": "npm run generate --workspace=@axonivy/ui-icons",
     "build": "lerna run build",
+    "build:storybook": "storybook build",
     "package": "lerna run package",
     "clean": "lerna run clean",
     "type": "lerna run type",
@@ -18,22 +19,36 @@
     "lint:ci": "eslint -o eslint.xml -f checkstyle",
     "lint:fix": "eslint --fix",
     "lint:inspect": "eslint --inspect-config",
-    "dev": "npm run dev --workspace=@axonivy/ui-components",
+    "dev": "storybook dev -p 6006 --no-open",
     "test": "vitest test",
     "test:ci": "vitest test --watch=false --reporter=default --reporter=junit --outputFile=report.xml",
     "publish:next": "lerna publish --exact --canary --preid next --pre-dist-tag next --no-git-tag-version --no-push --ignore-scripts --yes"
   },
   "devDependencies": {
+    "@storybook/addon-essentials": "^8.5.3",
+    "@storybook/addon-interactions": "^8.5.3",
+    "@storybook/addon-links": "^8.5.3",
+    "@storybook/blocks": "^8.5.3",
+    "@storybook/manager-api": "^8.5.3",
+    "@storybook/react": "^8.5.3",
+    "@storybook/react-vite": "^8.5.3",
+    "@storybook/test": "^8.5.3",
+    "@storybook/theming": "^8.5.3",
     "@types/node": "^22.10.7",
+    "@vanilla-extract/vite-plugin": "^5.0.0",
+    "@vitejs/plugin-react": "^4.3.4",
+    "eslint-plugin-storybook": "^0.12.0",
     "lerna": "^8.1.9",
     "rimraf": "^6.0.1",
     "rollup-plugin-visualizer": "^5.14.0",
+    "storybook": "^8.5.3",
     "typescript": "^5.7.3",
     "typescript-cp": "^0.1.9",
     "vite": "^6.0.0",
     "vite-plugin-dts": "^4.5.0",
     "vitest": "^3.0.2"
   },
+  "type": "module",
   "workspaces": {
     "packages": [
       "packages/*",
