+++ packages/components/package.json
@@ -41,28 +41,15 @@
     "react": "^18.2 || ^19.0"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "^8.5.3",
-    "@storybook/addon-interactions": "^8.5.3",
-    "@storybook/addon-links": "^8.5.3",
-    "@storybook/blocks": "^8.5.3",
-    "@storybook/manager-api": "^8.5.3",
-    "@storybook/react": "^8.5.3",
-    "@storybook/react-vite": "^8.5.3",
-    "@storybook/test": "^8.5.3",
-    "@storybook/theming": "^8.5.3",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^14.6.0",
     "@types/react": "^19.0.7",
     "@vanilla-extract/css": "^1.17.0",
     "@vanilla-extract/recipes": "^0.5.5",
-    "@vanilla-extract/vite-plugin": "^5.0.0",
-    "@vitejs/plugin-react": "^4.3.4",
-    "eslint-plugin-storybook": "^0.12.0",
     "happy-dom": "^17.0.0",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "storybook": "^8.5.3"
+    "react-dom": "^19.0.0"
   },
   "type": "module",
   "types": "./lib/index.d.ts",
@@ -72,8 +59,6 @@
     "clean": "rimraf lib *.tsbuildinfo",
     "build": "tsc --build tsconfig.production.json",
     "package": "npm run clean && vite build && npm run build",
-    "build:storybook": "storybook build",
-    "dev": "storybook dev -p 6006 --no-open",
     "type": "tsc --noEmit --emitDeclarationOnly false",
     "test": "vitest",
     "test:ci": "vitest --watch=false"
